# docs/openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecureVOTE API",
    "description": "Educational/research prototype of an embedded electronic-voting and election-integrity pipeline. This is NOT production election infrastructure. See /docs for the interactive API explorer.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/elections/{election_id}": {
      "get": {
        "tags": [
          "elections"
        ],
        "summary": "Get Election",
        "description": "Get election details with candidates and device count.",
        "operationId": "get_election_api_elections__election_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ElectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/elections/{election_id}/voters": {
      "post": {
        "tags": [
          "Voters"
        ],
        "summary": "Register Voter",
        "operationId": "register_voter_api_elections__election_id__voters_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoterCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Voters"
        ],
        "summary": "List Voters",
        "operationId": "list_voters_api_elections__election_id__voters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoterResponse"
                  },
                  "title": "Response List Voters Api Elections  Election Id  Voters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/voters/{voter_id}": {
      "get": {
        "tags": [
          "Voters"
        ],
        "summary": "Get Voter",
        "operationId": "get_voter_api_elections__election_id__voters__voter_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          },
          {
            "name": "voter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/eligibility": {
      "post": {
        "tags": [
          "Voters"
        ],
        "summary": "Check Eligibility",
        "operationId": "check_eligibility_api_elections__election_id__eligibility_post",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EligibilityCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EligibilityCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/elections/{election_id}/voter-lookup": {
      "get": {
        "tags": [
          "Voters"
        ],
        "summary": "Lookup Voter",
        "operationId": "lookup_voter_api_elections__election_id__voter_lookup_get",
        "parameters": [
          {
            "name": "election_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Election Id"
            }
          },
          {
            "name": "voter_id_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voter Id Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "party": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Party"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "name",
          "party",
          "symbol",
          "position",
          "created_at"
        ],
        "title": "CandidateResponse"
      },
...
      "ElectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "configuration_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Hash"
          },
          "total_ballots": {
            "type": "integer",
            "title": "Total Ballots"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "configured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured At"
          },
          "locked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked At"
          },
          "opened_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateResponse"
            },
            "type": "array",
            "title": "Candidates",
            "default": []
          },
          "device_count": {
            "type": "integer",
            "title": "Device Count",
            "default": 0
          },
          "ballot_count": {
            "type": "integer",
            "title": "Ballot Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "state",
          "configuration_hash",
          "total_ballots",
          "created_at",
          "configured_at",
          "locked_at",
          "opened_at",
          "closed_at",
          "published_at"
        ],
        "title": "ElectionResponse"
      },
...
      "EligibilityCheckRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "title": "Name"
          },
          "date_of_birth": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Date Of Birth"
          },
          "constituency": {
            "type": "string",
            "minLength": 2,
            "title": "Constituency"
          }
        },
        "type": "object",
        "required": [
          "name",
          "date_of_birth",
          "constituency"
        ],
        "title": "EligibilityCheckRequest"
      },
      "EligibilityCheckResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "voter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voter Id"
          },
          "notice": {
            "type": "string",
            "title": "Notice",
            "default": "SIMULATED ELIGIBILITY CHECK - NOT A REAL GOVERNMENT SERVICE"
          }
        },
        "type": "object",
        "required": [
          "status",
          "reasons"
        ],
        "title": "EligibilityCheckResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VoterCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Name"
          },
          "date_of_birth": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Date Of Birth"
          },
          "constituency": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Constituency"
          }
        },
        "type": "object",
        "required": [
          "name",
          "date_of_birth",
          "constituency"
        ],
        "title": "VoterCreate"
      },
      "VoterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "voter_id_number": {
            "type": "string",
            "title": "Voter Id Number"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date_of_birth": {
            "type": "string",
            "title": "Date Of Birth"
          },
          "constituency": {
            "type": "string",
            "title": "Constituency"
          },
          "polling_station_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polling Station Id"
          },
          "eligibility_status": {
            "type": "string",
            "title": "Eligibility Status"
          },
          "registration_status": {
            "type": "string",
            "title": "Registration Status"
          },
          "has_voted": {
            "type": "boolean",
            "title": "Has Voted"
          },
          "registered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Registered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "election_id",
          "voter_id_number",
          "name",
          "date_of_birth",
          "constituency",
          "polling_station_id",
          "eligibility_status",
          "registration_status",
          "has_voted",
          "registered_at"
        ],
        "title": "VoterResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}